+++ src/common/components/index.js
@@ -2,6 +2,7 @@ export { default as AddressInfo } from './AddressInfo';
 export { default as ButtonIsPrimary } from './ButtonIsPrimary';
 export { Credentials } from './Credentials';
 export { CredentialsField } from './CredentialsField';
+export { Ellipsis } from './Ellipsis';
 export { default as EntitiesWithCollapsing } from './EntitiesWithCollapsing';
 export { default as FieldIsPrimary } from './FieldIsPrimary';
 export { default as FieldLanguage } from './FieldLanguage';