+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.1.0 (IN PROGRESS)
 
 * Unpin `@rehooks/local-storage` now that it's no longer broken. Refs UIORGS-344.
+* Add `Note` field to contact people summary list. Refs UIORGS-352.
 
 ## [4.0.0](https://github.com/folio-org/ui-organizations/tree/v4.0.0) (2023-02-22)
 [Full Changelog](https://github.com/folio-org/ui-organizations/compare/v3.3.1...v4.0.0)